+++ routes/api.php
@@ -20,14 +20,12 @@ Route::group([
 //endregion
 
 //region Category
-Route::get('/category/{id?}', [CategoryController::class, 'index']);
-
 Route::group([
     'middleware' => 'api',
     'prefix' => 'category'
 ], function ($router) {
-
-    Route::put('/', [CategoryController::class, 'store']);
+    Route::get('/category/{id?}', [CategoryController::class, 'index']);
+    Route::post('/', [CategoryController::class, 'store']);
     Route::delete('/', [CategoryController::class, 'destroy']);
 });
 
@@ -39,7 +37,7 @@ Route::group([
     'prefix' => 'lists'
 ], function ($router) {
     Route::get('/{id?}', [ListController::class, 'index']);
-    Route::put('/', [ListController::class, 'store']);
+    Route::post('/', [ListController::class, 'store']);
     Route::delete('/{id}', [ListController::class, 'destroy']);
 });
 //endregion
@@ -50,7 +48,7 @@ Route::group([
     'prefix' => 'items'
 ], function ($router) {
     Route::get('/{listId?}/{itemId?}', [ItemController::class, 'index']);
-    Route::put('/', [ItemController::class, 'store']);
+    Route::post('/', [ItemController::class, 'store']);
     Route::patch('/{id}', [ItemController::class, 'update']);
     Route::delete('/', [ItemController::class, 'destroy']);
 });